requirements.txt
streamlit
streamlit_optionmenu
numpy 
tensorflow
pillow
keras